refactor(pipeline): split monolithic pipeline.py into src/pipeline/ package
