--- packages/backend/src/pipeline.py
@@ -1,1773 +1,0 @@
-"""
-Enterprise Policy Document Crawling Pipeline
-
-This module provides a comprehensive crawling pipeline that integrates the ClauseaCrawler
-with AI-powered document analysis for policy document discovery and processing.
-
-Architecture Decisions:
-1. **Modular Design**: Separates concerns into distinct classes for crawling, analysis,
-   and storage, enabling easy testing and maintenance.
-
-2. **Async/Await Pattern**: Uses asyncio throughout for optimal I/O performance when
-   dealing with multiple network requests and database operations.
-
-3. **Comprehensive Error Handling**: Implements graceful error handling with detailed
-   logging to ensure pipeline robustness and easy debugging.
-
-4. **Memory Management**: Includes memory monitoring and optimization strategies for
-   large-scale crawling operations.
-
-5. **Rate Limiting**: Built-in respect for robots.txt and configurable delays to be
-   a good web citizen.
-
-6. **Deduplication**: Smart URL and content deduplication to avoid processing
-   duplicate documents.
-
-7. **Incremental Processing**: Processes products sequentially to manage memory
-   and respect rate limits while maintaining data consistency.
-
-Performance Characteristics:
-- Memory efficient: Processes products one at a time
-- Network optimized: Concurrent requests with configurable limits
-- Database efficient: Bulk operations and smart update logic
-- Scalable: Can handle hundreds of products and thousands of documents
-
-Usage:
-    # Run the complete pipeline
-    python -m src.pipeline
-
-    # Or use programmatically
-    from src.pipeline import PolicyDocumentPipeline
-
-    pipeline = PolicyDocumentPipeline()
-    results = await pipeline.run()
-"""
-
-import asyncio
-import hashlib
-import json
-import os
-import re
-import time
-import tracemalloc
-from collections.abc import Awaitable, Callable
-from datetime import datetime, timedelta
-from typing import Any
-from urllib.parse import urlparse
-
-import tldextract
-from dotenv import load_dotenv
-from pydantic import BaseModel, Field
-
-from src.analyzers.date_extractor import DateExtractor
-from src.analyzers.document_classifier import DocumentClassifier
-from src.analyzers.locale_analyzer import LocaleAnalyzer
-from src.analyzers.region_detector import RegionDetector
-from src.core.database import db_session
-from src.core.logging import get_logger
-from src.crawler import ClauseaCrawler, CrawlResult
-from src.llm import SupportedModel, acompletion_with_fallback
-from src.models.crawl import CrawlSession
-from src.models.document import Document, coerce_doc_type_from_classifier
-from src.models.pipeline_job import classify_crawl_error
-from src.models.product import Product
-from src.repositories.crawl_repository import CrawlRepository
-from src.repositories.document_version_repository import DocumentVersionRepository
-from src.services.service_factory import create_document_service, create_product_service
-from src.utils.llm_usage import usage_tracking
-from src.utils.llm_usage_tracking_mixin import LLMUsageTrackingMixin
-from src.utils.markdown import markdown_to_text
-from src.utils.perf import log_memory_usage, memory_monitor_task
-
-load_dotenv()
-
-logger = get_logger(__name__)
-
-# Offline extractor (no live suffix-list fetch), matching the crawler's domain matching.
-_TLD_EXTRACT = tldextract.TLDExtract(suffix_list_urls=())
-
-logger_discovery = get_logger(__name__, component="pipeline:discovery")
-logger_analysis = get_logger(__name__, component="pipeline:analysis")
-logger_storage = get_logger(__name__, component="pipeline:storage")
-
-# Minimum normalized legal_score (0.0–1.0) to accept a crawled page into analysis.
-MIN_LEGAL_SCORE_THRESHOLD = 0.2
-
-# Resume freshness window (hours). A retried crawl skips re-fetching docs stored
-# within this window so it resumes instead of re-rendering hours of work. Stored
-# older than this -> still re-fetched, so scheduled monitoring re-crawls days later
-# continue to detect policy changes. Recency is the signal that tells a retry
-# (minutes/hours later) apart from a monitoring run (days later).
-RESUME_FRESH_HOURS = float(os.getenv("PIPELINE_RESUME_FRESH_HOURS", "24"))
-
-
-def _content_fingerprint(text: str) -> str:
-    """Generate a fingerprint for near-duplicate detection.
-
-    Normalizes whitespace and lowercases before hashing to catch
-    slight formatting differences between URLs serving the same content.
-    """
-    normalized = re.sub(r"\s+", " ", text.lower().strip())
-    return hashlib.md5(normalized[:5000].encode()).hexdigest()
-
-
-def _diff_fields(existing: "Document", incoming: "Document") -> list[str]:
-    tracked = ["text", "title", "doc_type", "locale", "regions", "effective_date"]
-    changed = []
-    for field in tracked:
-        old_val = getattr(existing, field)
-        new_val = getattr(incoming, field)
-        if field == "regions":
-            if set(old_val or []) != set(new_val or []):
-                changed.append(field)
-        elif old_val != new_val:
-            changed.append(field)
-    return changed
-
-
-# Locale path segment (/es/, /fr/, /pt-br/, /en_US/) or locale subdomain (es., de.).
-# Used only to prefer a canonical URL when collapsing same-content duplicates — the
-# collapse decision itself is content-based, so genuinely different regional policies
-# (different text) are never merged.
-_LOCALE_PATH_RE = re.compile(r"^/[a-z]{2}([-_][a-z]{2})?(/|$)", re.IGNORECASE)
-_LOCALE_HOST_RE = re.compile(r"^[a-z]{2}([-_][a-z]{2})?\.", re.IGNORECASE)
-
-
-def _canonical_rank(url: str) -> tuple[int, int]:
-    """Lower rank = more canonical. Prefers non-locale URLs, then shorter paths.
-
-    Ensures that when several URLs serve identical content (e.g. /privacy and
-    /es/privacy), the locale-neutral one is kept as the representative.
-    """
-    parsed = urlparse(url)
-    host = parsed.netloc.lower()
-    path = parsed.path or "/"
-    looks_locale = bool(_LOCALE_PATH_RE.match(path)) or (
-        bool(_LOCALE_HOST_RE.match(host)) and not host.startswith("www.")
-    )
-    return (1 if looks_locale else 0, len(path))
-
-
-class ProcessingStats(BaseModel):
-    """Statistics for document processing pipeline."""
-
-    products_processed: int = 0
-    products_failed: int = 0
-    failed_product_slugs: list[str] = Field(default_factory=list)
-    total_urls_crawled: int = 0
-    total_documents_found: int = 0
-    policy_documents_processed: int = 0
-    policy_documents_stored: int = 0
-    english_documents: int = 0
-    non_english_skipped: int = 0
-    duplicates_skipped: int = 0
-    processing_time_seconds: float = 0.0
-    total_prompt_tokens: int = 0
-    total_completion_tokens: int = 0
-    total_tokens: int = 0
-    total_cost: float = 0.0
-
-    # Per-URL crawl failures collected during the pipeline run
-    crawl_errors: list[dict[str, Any]] = Field(default_factory=list)
-
-    # Per-URL silent skips — pages fetched OK but rejected by a filter
-    # (e.g. insufficient_content, low_legal_score, non_policy_classification).
-    # Mirrors models.pipeline_job.CrawlSkip; kept as dicts here to avoid an
-    # import cycle.
-    crawl_skip_reasons: list[dict[str, Any]] = Field(default_factory=list)
-
-    @property
-    def success_rate(self) -> float:
-        """Calculate product processing success rate."""
-        total = self.products_processed + self.products_failed
-        return (self.products_processed / total * 100) if total > 0 else 0.0
-
-    @property
-    def legal_detection_rate(self) -> float:
-        """Calculate policy document detection rate."""
-        return (
-            (self.policy_documents_processed / self.total_documents_found * 100)
-            if self.total_documents_found > 0
-            else 0.0
-        )
-
-
-class DocumentAnalyzer(LLMUsageTrackingMixin):
-    """
-    AI-powered document analyzer for locale detection, legal classification, and region analysis.
-
-    This class encapsulates all AI/LLM interactions for document analysis, providing
-    a clean interface for the main pipeline while handling API errors gracefully.
-    """
-
-    def __init__(
-        self,
-        model_name: SupportedModel | None = None,
-        max_content_length: int = 5000,
-    ):
-        """
-        Initialize the document analyzer.
-
-        Args:
-            model_name: Optional LLM model to use for analysis. If None, uses default priority list with fallback.
-            max_content_length: Maximum content length to send to LLM
-        """
-        super().__init__()
-
-        self.model_name = model_name
-        self.max_content_length = max_content_length
-
-        # Initialize specialized analyzers
-
-        self.locale_analyzer = LocaleAnalyzer()
-        self.document_classifier = DocumentClassifier(max_content_length=max_content_length)
-        self.region_detector = RegionDetector()
-        self.date_extractor = DateExtractor()
-
-    def reset_usage_stats(self) -> None:
-        """Clear recorded LLM usage statistics across all analyzers."""
-        super().reset_usage_stats()
-        self.locale_analyzer.reset_usage_stats()
-        self.document_classifier.reset_usage_stats()
-        self.date_extractor.reset_usage_stats()
-
-    def get_usage_summary(self) -> dict[str, dict[str, Any]]:
-        """Aggregate token usage from all analyzers without clearing records."""
-        summary = super().get_usage_summary()
-        for analyzer in (self.locale_analyzer, self.document_classifier, self.date_extractor):
-            summary = self._merge_usage_summary(summary, analyzer.get_usage_summary())
-        return summary
-
-    def consume_usage_summary(self) -> tuple[dict[str, dict[str, Any]], list[Any]]:
-        """Return aggregated usage information and clear recorded statistics."""
-        summary, records = super().consume_usage_summary()
-        for analyzer in (self.locale_analyzer, self.document_classifier, self.date_extractor):
-            analyzer_summary, analyzer_records = analyzer.consume_usage_summary()
-            summary = self._merge_usage_summary(summary, analyzer_summary)
-            records.extend(analyzer_records)
-        return summary, records
-
-    def _merge_usage_summary(
-        self, base: dict[str, dict[str, Any]], incoming: dict[str, dict[str, Any]]
-    ) -> dict[str, dict[str, Any]]:
-        for model, stats in incoming.items():
-            entry = base.setdefault(
-                model,
-                {
-                    "prompt_tokens": 0,
-                    "completion_tokens": 0,
-                    "total_tokens": 0,
-                    "cost": 0.0,
-                    "provider_models": [],
-                },
-            )
-            entry["prompt_tokens"] += stats.get("prompt_tokens", 0)
-            entry["completion_tokens"] += stats.get("completion_tokens", 0)
-            entry["total_tokens"] += stats.get("total_tokens", 0)
-            if stats.get("cost"):
-                entry["cost"] = (entry.get("cost", 0.0) or 0.0) + stats.get("cost", 0.0)
-
-            providers = set(entry.get("provider_models") or [])
-            providers.update(stats.get("provider_models") or [])
-            entry["provider_models"] = sorted(providers)
-
-        return base
-
-    async def detect_locale(
-        self, text: str, metadata: dict[str, Any], url: str | None = None
-    ) -> dict[str, Any]:
-        """
-        Detect the locale of a document.
-
-        Delegates to the specialized LocaleAnalyzer.
-
-        Args:
-            text: Document content
-            metadata: Document metadata
-            url: Optional document URL for pattern analysis
-
-        Returns:
-            Dict containing locale, confidence, and language_name
-        """
-        return await self.locale_analyzer.detect_locale(text, metadata, url)
-
-    async def classify_document(
-        self, url: str, text: str, metadata: dict[str, Any]
-    ) -> dict[str, Any]:
-        """
-        Classify if document is a policy document and determine its type.
-
-        Delegates to the specialized DocumentClassifier.
-
-        Args:
-            url: Document URL
-            text: Document content
-            metadata: Document metadata
-
-        Returns:
-            Dict containing classification, justification, and is_policy_document flag
-        """
-        return await self.document_classifier.classify_document(url, text, metadata)
-
-    async def detect_regions(self, text: str, metadata: dict[str, Any], url: str) -> dict[str, Any]:
-        """
-        Detect if document applies globally or to specific regions.
-
-        Delegates to the specialized RegionDetector.
-
-        Args:
-            text: Document content
-            metadata: Document metadata
-            url: Document URL
-
-        Returns:
-            Dict containing region analysis with mapped region codes
-        """
-        return await self.region_detector.detect_regions(text, metadata, url)
-
-    async def extract_title(
-        self, markdown: str, metadata: dict[str, Any], url: str, doc_type: str
-    ) -> dict[str, Any]:
-        """
-        Extract meaningful title from document.
-
-        Args:
-            markdown: Document markdown content
-            metadata: Document metadata
-            url: Document URL
-            doc_type: Classified document type
-
-        Returns:
-            Dict containing extracted title and confidence
-        """
-        # Quick extraction from metadata first
-        if metadata:
-            for key in ["title", "og:title"]:
-                if key in metadata and metadata[key]:
-                    title = metadata[key].strip()
-                    if title:
-                        return {"title": title, "confidence": 0.9}
-
-        # Extract from markdown content
-        lines = markdown.split("\n")
-        for line in lines[:10]:  # Check first 10 lines
-            line = line.strip()
-            if line.startswith("#") and len(line) < 200:
-                title = line.lstrip("#").strip()
-                if title:
-                    return {"title": title, "confidence": 0.8}
-
-        # Fallback to document type with domain
-        from urllib.parse import urlparse
-
-        domain = urlparse(url).netloc.replace("www.", "")
-        type_titles = {
-            "privacy_policy": "Privacy Policy",
-            "terms_of_service": "Terms of Service",
-            "cookie_policy": "Cookie Policy",
-            "terms_and_conditions": "Terms and Conditions",
-            "data_processing_agreement": "Data Processing Agreement",
-            "gdpr_policy": "GDPR Policy",
-            "copyright_policy": "Copyright Policy",
-            "community_guidelines": "Community Guidelines",
-            "children_privacy_policy": "Children's Privacy Policy",
-            "security_policy": "Security Practices",
-        }
-
-        title = f"{type_titles.get(doc_type, 'Policy Document')} - {domain}"
-        return {"title": title, "confidence": 0.5}
-
-    async def extract_effective_date(self, content: str, metadata: dict[str, Any]) -> str | None:
-        """
-        Extract the effective date from a policy document.
-
-        Args:
-            content: Document text content
-            metadata: Document metadata
-
-        Returns:
-            Effective date as ISO string (YYYY-MM-DD) or None if not found
-        """
-        return await self.date_extractor.extract_effective_date(content, metadata)
-
-    async def _extract_effective_date_static(
-        self, content: str, metadata: dict[str, Any]
-    ) -> str | None:
-        """
-        Attempt static extraction of effective date from metadata and content patterns.
-
-        Args:
-            content: Document text content
-            metadata: Document metadata
-
-        Returns:
-            Effective date as ISO string or None
-        """
-        import re
-
-        # Check metadata first
-        if metadata:
-            for key in ["effective_date", "last_updated", "date", "published"]:
-                if key in metadata and metadata[key]:
-                    date_str = str(metadata[key]).strip()
-                    parsed_date = self._parse_date_string(date_str)
-                    if parsed_date:
-                        return parsed_date
-
-        # Common effective date patterns in policy documents (ordered by specificity)
-        # More specific patterns first to avoid false matches
-        patterns = [
-            # Explicit "Effective date:" or "Effective as of:"
-            r"effective\s+(?:date|as\s+of):\s*([^.\n<]+)",
-            # "Last updated:" or "Last modified:"
-            r"last\s+(?:updated|modified|revised|reviewed):?\s*([^.\n<]+)",
-            # "Updated on:" or "Modified on:"
-            r"(?:updated|modified|revised)\s+on:?\s*([^.\n<]+)",
-            # "Revision date:" or "Version date:"
-            r"(?:revision|version)\s+date:?\s*([^.\n<]+)",
-            # "This policy is effective..." or "This agreement takes effect..."
-            r"(?:this\s+(?:policy|agreement|document|terms?|privacy|cookie)\s+)?(?:is\s+)?(?:effective|takes\s+effect|became\s+effective)\s+(?:as\s+of\s+)?([^.\n<]+)",
-            # "Effective:" standalone
-            r"effective:?\s*([^.\n<]+)",
-            # "Date of effect:" or "Date effective:"
-            r"date\s+(?:of\s+effect|effective):?\s*([^.\n<]+)",
-            # "Published:" or "Publication date:"
-            r"(?:publication|publish)\s+date:?\s*([^.\n<]+)",
-            # "Posted:" or "Posted on:"
-            r"(?:posted|post)\s+(?:on|date)?:?\s*([^.\n<]+)",
-            # "Created:" or "Creation date:"
-            r"(?:creation|created?)\s+date:?\s*([^.\n<]+)",
-            # ISO date format in context (YYYY-MM-DD)
-            r"(?:effective|updated|modified|published|posted|created).*?(\d{4}[-/]\d{1,2}[-/]\d{1,2})",
-            # US/European numeric date format (MM/DD/YYYY or DD/MM/YYYY)
-            r"(?:effective|updated|modified|published).*?(\d{1,2}[/-]\d{1,2}[/-]\d{4})",
-            # Full month names
-            r"(?:effective|updated|modified|published).*?((?:january|february|march|april|may|june|july|august|september|october|november|december)\s+\d{1,2}(?:st|nd|rd|th)?,?\s+\d{4})",
-            # Abbreviated month names
-            r"(?:effective|updated).*?((?:jan|feb|mar|apr|may|jun|jul|aug|sep|oct|nov|dec)[\s\.]\d{1,2}(?:st|nd|rd|th)?,?\s+\d{4})",
-            # Relative dates (for fallback, but try to convert to actual dates)
-            r"effective\s+immediately",
-            r"effective\s+upon\s+(?:publication|posting|acceptance)",
-            r"effective\s+(?:from|as\s+of)\s+(?:the\s+)?date\s+(?:of|this)",
-            # Date ranges (take the start date)
-            r"effective\s+(?:from|between)\s+([^.\n<]+?)(?:\s+(?:to|through|until)\s+[^.\n<]+)",
-            r"effective\s+([^.\n<]+?)\s+through\s+([^.\n<]+)",  # Capture first date
-            # International date formats
-            r"(?:effective|updated).*?(\d{1,2}\.\d{1,2}\.\d{4})",  # DD.MM.YYYY (German)
-            r"(?:effective|updated).*?(\d{4}年\d{1,2}月\d{1,2}日)",  # YYYY年MM月DD日 (Japanese)
-            # Ordinal indicators
-            r"(?:effective|updated).*?(first|1st|second|2nd|third|3rd|fourth|4th|fifth|5th|sixth|6th|seventh|7th|eighth|8th|ninth|9th|tenth|10th)\s+(?:of\s+)?(?:january|february|march|april|may|june|july|august|september|october|november|december),?\s+\d{4}",
-        ]
-
-        # Search in first 5000 chars where dates are typically mentioned
-        search_text = content[:5000].lower()
-
-        for pattern in patterns:
-            matches = re.finditer(pattern, search_text, re.IGNORECASE | re.MULTILINE)
-            for match in matches:
-                date_str = match.group(1).strip()
-                # Clean up common trailing words/phrases
-                date_str = re.sub(r"\s*(and|or|,|;|\.|$).*$", "", date_str, flags=re.IGNORECASE)
-                date_str = date_str.strip()
-
-                if date_str:
-                    parsed_date = self._parse_date_string(date_str)
-                    if parsed_date:
-                        logger.debug(
-                            f"Extracted date from pattern '{pattern}': {date_str} -> {parsed_date}"
-                        )
-                        return parsed_date
-
-        return None
-
-    async def _extract_effective_date_llm(
-        self, content: str, metadata: dict[str, Any]
-    ) -> str | None:
-        """
-        Use LLM to extract effective date from document content.
-
-        Args:
-            content: Document text content
-            metadata: Document metadata
-
-        Returns:
-            Effective date as ISO string or None if not found
-        """
-        # Use first portion of content where dates are typically mentioned
-        content_sample = content[:3000] if len(content) > 3000 else content
-
-        prompt = f"""Analyze this policy document to find the effective date.
-
-Content: {content_sample}
-Metadata: {json.dumps(metadata, indent=2) if metadata else "None"}
-
-Look for:
-- "Effective date:", "Effective as of:", etc.
-- "Last updated:", "Updated on:", etc.
-- "This policy is effective...", "This agreement takes effect..."
-- Any explicit date mentioned as when the document becomes effective
-
-Return JSON:
-{{
-    "effective_date": "YYYY-MM-DD" or null,
-    "confidence": float 0-1,
-    "source_text": "exact text snippet where date was found" or null,
-    "reasoning": "explanation of why this date was chosen or why none found"
-}}
-
-IMPORTANT: Return null for effective_date if you cannot find a clear effective date. Do not guess or infer dates."""
-
-        system_prompt = """You are an expert at extracting effective dates from policy documents. Only return dates that are explicitly stated as effective dates, last updated dates, or similar. Do not guess or infer dates."""
-
-        try:
-            async with usage_tracking(self._create_usage_tracker("extract_effective_date")):
-                response = await acompletion_with_fallback(
-                    messages=[
-                        {"role": "system", "content": system_prompt},
-                        {"role": "user", "content": prompt},
-                    ],
-                    response_format={"type": "json_object"},
-                )
-
-            choice = response.choices[0]
-            if not hasattr(choice, "message"):
-                raise ValueError("Unexpected response format: missing message attribute")
-            message = choice.message  # type: ignore[attr-defined]
-            if not message:
-                raise ValueError("Unexpected response format: message is None")
-            content = message.content  # type: ignore
-            if not content:
-                raise ValueError("Empty response from LLM")
-
-            result = json.loads(content)
-            effective_date = result.get("effective_date")
-
-            if effective_date:
-                # Validate the date format
-                parsed_date = self._parse_date_string(effective_date)
-                if parsed_date:
-                    logger.debug(
-                        f"LLM found effective date: {parsed_date} "
-                        f"(confidence: {result.get('confidence', 0):.2f})"
-                    )
-                    return parsed_date
-
-            logger.debug("LLM could not find effective date")
-            return None
-
-        except Exception as e:
-            logger.warning(f"LLM effective date extraction failed: {e}")
-            return None
-
-    def _parse_date_string(self, date_str: str) -> str | None:
-        """
-        Parse a date string into ISO format (YYYY-MM-DD).
-
-        Enhanced to handle more formats, ordinal indicators, and international variations.
-
-        Args:
-            date_str: Date string to parse
-
-        Returns:
-            ISO formatted date string or None if parsing fails
-        """
-        if not date_str or not isinstance(date_str, str):
-            return None
-
-        # Clean up the date string
-        date_str = date_str.strip().replace(",", "").replace(".", "").lower()
-
-        # Handle relative dates (return None as they're not specific dates)
-        if any(
-            word in date_str for word in ["immediately", "upon", "as of the date", "as of this"]
-        ):
-            return None
-
-        # Handle ordinal indicators (convert to cardinal numbers)
-        ordinals = {
-            "first": "1",
-            "1st": "1",
-            "second": "2",
-            "2nd": "2",
-            "third": "3",
-            "3rd": "3",
-            "fourth": "4",
-            "4th": "4",
-            "fifth": "5",
-            "5th": "5",
-            "sixth": "6",
-            "6th": "6",
-            "seventh": "7",
-            "7th": "7",
-            "eighth": "8",
-            "8th": "8",
-            "ninth": "9",
-            "9th": "9",
-            "tenth": "10",
-            "10th": "10",
-        }
-        for ordinal, number in ordinals.items():
-            date_str = date_str.replace(ordinal, number)
-
-        # Extended date formats to try
-        formats = [
-            # ISO and standard formats
-            "%Y-%m-%d",  # 2023-12-01
-            "%Y/%m/%d",  # 2023/12/01
-            "%m/%d/%Y",  # 12/01/2023
-            "%d/%m/%Y",  # 01/12/2023
-            "%d.%m.%Y",  # 01.12.2023 (German style)
-            # Full month names
-            "%B %d %Y",  # December 1 2023
-            "%b %d %Y",  # Dec 1 2023
-            "%d %B %Y",  # 1 December 2023
-            "%d %b %Y",  # 1 Dec 2023
-            # With commas and periods
-            "%B %d, %Y",  # December 1, 2023
-            "%b %d, %Y",  # Dec 1, 2023
-            # ISO with time components
-            "%Y-%m-%dT%H:%M:%S",  # ISO with time
-            "%Y-%m-%d %H:%M:%S",  # Standard with time
-            "%Y-%m-%dT%H:%M:%SZ",  # ISO with timezone
-            # Japanese format
-            "%Y年%m月%d日",  # 2023年12月1日
-        ]
-
-        for fmt in formats:
-            try:
-                parsed = datetime.strptime(date_str, fmt)
-                return parsed.strftime("%Y-%m-%d")
-            except ValueError:
-                continue
-
-        # Try to extract various year-month-day patterns with regex
-        patterns = [
-            r"(\d{4})[/-](\d{1,2})[/-](\d{1,2})",  # YYYY-MM-DD or YYYY/MM/DD
-            r"(\d{1,2})[/-](\d{1,2})[/-](\d{4})",  # MM/DD/YYYY or DD/MM/YYYY (ambiguous)
-        ]
-
-        for pattern in patterns:
-            match = re.search(pattern, date_str)
-            if match:
-                groups = match.groups()
-                if len(groups[0]) == 4:  # YYYY-MM-DD format
-                    year, month, day = groups
-                else:  # Assume MM/DD/YYYY for US-centric parsing
-                    month, day, year = groups
-
-                try:
-                    parsed = datetime(int(year), int(month), int(day))
-                    return parsed.strftime("%Y-%m-%d")
-                except ValueError:
-                    continue
-
-        # Try to parse month name + day + year
-        month_names = {
-            "january": 1,
-            "february": 2,
-            "march": 3,
-            "april": 4,
-            "may": 5,
-            "june": 6,
-            "july": 7,
-            "august": 8,
-            "september": 9,
-            "october": 10,
-            "november": 11,
-            "december": 12,
-            "jan": 1,
-            "feb": 2,
-            "mar": 3,
-            "apr": 4,
-            "jun": 6,
-            "jul": 7,
-            "aug": 8,
-            "sep": 9,
-            "oct": 10,
-            "nov": 11,
-            "dec": 12,
-        }
-
-        # Pattern: month day, year or month day year
-        month_day_year_pattern = (
-            r"(" + "|".join(month_names.keys()) + r")\s+(\d{1,2})(?:st|nd|rd|th)?,?\s+(\d{4})"
-        )
-        match = re.search(month_day_year_pattern, date_str, re.IGNORECASE)
-        if match:
-            month_name, day, year = match.groups()
-            month = month_names[month_name.lower()]
-            try:
-                parsed = datetime(int(year), month, int(day))
-                return parsed.strftime("%Y-%m-%d")
-            except ValueError:
-                pass
-
-        return None
-
-
-class PolicyDocumentPipeline:
-    """
-    Main pipeline orchestrator for policy document crawling and processing.
-
-    This class coordinates the entire pipeline from product retrieval to document storage,
-    providing comprehensive error handling, logging, and performance monitoring.
-    """
-
-    def __init__(
-        self,
-        max_depth: int | None = None,
-        max_pages: int | None = None,
-        crawler_strategy: str | None = None,
-        concurrent_limit: int | None = None,
-        delay_between_requests: float | None = None,
-        timeout: int | None = None,
-        respect_robots_txt: bool | None = None,
-        max_parallel_products: int | None = None,
-        use_browser: bool | None = None,
-        proxy: str | None = None,
-        fallback_min_legal_score: float | None = None,
-        discovery_min_legal_score: float | None = None,
-        discovery_strategy: str | None = None,
-        fallback_strategy: str | None = None,
-        min_docs_before_fallback: int | None = None,
-        required_doc_types: list[str] | None = None,
-        user_agent: str | None = None,
-        progress_callback: Callable[[str, int, int], None]
-        | Callable[[str, int, int], Awaitable[None]]
-        | None = None,  # Optional progress callback (phase, current, total)
-        job_id: str | None = None,
-    ):
-        """
-        Initialize the policy document pipeline.
-
-        Args:
-            max_depth: Maximum crawl depth per product (default: ``CrawlerConfig`` / env)
-            max_pages: Maximum pages to crawl per product
-            crawler_strategy: Default ClauseaCrawler strategy when not overridden per pass
-            concurrent_limit: Maximum concurrent requests
-            delay_between_requests: Delay between requests in seconds
-            timeout: Request timeout in seconds
-            respect_robots_txt: Whether to respect robots.txt
-            max_parallel_products: Maximum number of products to process in parallel
-            use_browser: Whether to use browser for crawling
-            proxy: Optional proxy URL (overrides ``CRAWLER_PROXY`` when set)
-            fallback_min_legal_score: Min legal score for fallback (mainly for best_first)
-            discovery_min_legal_score: Min legal URL score for discovery pass
-            discovery_strategy: Crawl strategy for discovery pass (e.g. best_first)
-            fallback_strategy: Crawl strategy for fallback pass (e.g. bfs)
-            min_docs_before_fallback: Minimum classified policy docs before skipping fallback
-            required_doc_types: Doc types that must be present to skip fallback
-            user_agent: HTTP User-Agent for the crawler
-        """
-        from src.core.config import config, discovery_crawl_limits
-
-        c = config.crawler
-
-        self.max_depth = max_depth if max_depth is not None else c.max_depth
-        self.max_pages = max_pages if max_pages is not None else c.max_pages
-        self.discovery_max_pages, self.discovery_max_depth = discovery_crawl_limits(
-            self.max_pages, self.max_depth
-        )
-        self.crawler_strategy = (
-            crawler_strategy if crawler_strategy is not None else c.crawler_strategy
-        )
-        self.concurrent_limit = (
-            concurrent_limit if concurrent_limit is not None else c.concurrent_limit
-        )
-        self.delay_between_requests = (
-            delay_between_requests
-            if delay_between_requests is not None
-            else c.delay_between_requests
-        )
-        self.timeout = timeout if timeout is not None else c.timeout
-        self.respect_robots_txt = (
-            respect_robots_txt if respect_robots_txt is not None else c.respect_robots_txt
-        )
-        self.max_parallel_products = (
-            max_parallel_products if max_parallel_products is not None else c.max_parallel_products
-        )
-        self.use_browser = use_browser if use_browser is not None else c.use_browser
-        self.browser_concurrency = c.browser_concurrency
-        self.proxy = proxy if proxy is not None else c.proxy
-        self.fallback_min_legal_score = (
-            fallback_min_legal_score
-            if fallback_min_legal_score is not None
-            else c.fallback_min_legal_score
-        )
-        self.discovery_min_legal_score = (
-            discovery_min_legal_score
-            if discovery_min_legal_score is not None
-            else c.discovery_min_legal_score
-        )
-        self.discovery_strategy = (
-            discovery_strategy if discovery_strategy is not None else c.discovery_strategy
-        )
-        self.fallback_strategy = (
-            fallback_strategy if fallback_strategy is not None else c.fallback_strategy
-        )
-        self.min_docs_before_fallback = (
-            min_docs_before_fallback
-            if min_docs_before_fallback is not None
-            else c.min_docs_before_fallback
-        )
-        self.required_doc_types = (
-            required_doc_types if required_doc_types is not None else list(c.required_doc_types)
-        )
-        self.user_agent = user_agent if user_agent is not None else c.user_agent
-        self.progress_callback: (
-            Callable[[str, int, int], None] | Callable[[str, int, int], Awaitable[None]] | None
-        ) = progress_callback
-        self._pending_progress_tasks: list[asyncio.Task] = []
-        self._job_id = job_id
-
-        # Initialize components
-        self.analyzer = DocumentAnalyzer()
-        self.stats = ProcessingStats()
-
-        logger.info(
-            "Pipeline initialized: max_depth=%s max_pages=%s discovery=%s/%s (pages/depth) "
-            "discovery_strategy=%s fallback_strategy=%s",
-            self.max_depth,
-            self.max_pages,
-            self.discovery_max_pages,
-            self.discovery_max_depth,
-            self.discovery_strategy,
-            self.fallback_strategy,
-        )
-
-    async def _get_recently_stored_urls(self, product: Product) -> list[str]:
-        """URLs of this product's docs stored within the resume freshness window.
-
-        Drives crawl resume: passed to the crawler so a retried long crawl skips
-        re-fetching pages it already stored minutes/hours ago. Docs older than
-        ``RESUME_FRESH_HOURS`` are excluded, so a scheduled monitoring re-crawl days
-        later still re-fetches them and detects policy changes. One query per crawl.
-        Failures are non-fatal — the crawl simply runs without the skip-set.
-        """
-        cutoff = datetime.now() - timedelta(hours=RESUME_FRESH_HOURS)
-        try:
-            async with db_session() as db:
-                document_service = create_document_service()
-                return await document_service.get_recent_document_urls(db, product.id, cutoff)
-        except Exception as error:
-            logger.warning(
-                f"Could not load recently stored URLs for {product.name} "
-                f"(resume skip disabled this run): {error}"
-            )
-            return []
-
-    @staticmethod
-    def _allowed_domains_for_product(product: Product) -> list[str]:
-        """Product domains unioned with the registered domain of every crawl seed.
-
-        A seed on a different registered domain (e.g. www.sheingroup.com seeded for a product
-        whose domain is shein.com) is otherwise rejected by the domain gate, dropping a whole
-        policy source. Deduped by registered-domain name, matching is_allowed_domain.
-        """
-        domains = list(product.domains or [])
-        seen = {_TLD_EXTRACT(d).domain for d in domains}
-        for seed in product.crawl_base_urls or []:
-            ext = _TLD_EXTRACT(seed)
-            if ext.domain and ext.domain not in seen:
-                seen.add(ext.domain)
-                domains.append(f"{ext.domain}.{ext.suffix}" if ext.suffix else ext.domain)
-        return domains
-
-    def _create_crawler_for_product(
-        self,
-        product: Product,
-        *,
-        max_depth: int | None = None,
-        max_pages: int | None = None,
-        min_legal_score: float | None = None,
-        strategy: str | None = None,
-        progress_phase: str | None = None,
-        result_callback: Callable[[CrawlResult], Awaitable[None]] | None = None,
-        stop_callback: Callable[[], bool] | Callable[[], Awaitable[bool]] | None = None,
-        recently_stored_urls: list[str] | None = None,
-    ) -> ClauseaCrawler:
-        """Create a configured crawler instance for a specific product."""
-        from datetime import datetime
-
-        timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-        log_filename = f"{timestamp}_{product.slug}_crawl.log"
-        log_file_path = f"logs/{log_filename}"
-
-        # Create progress callback that forwards to pipeline callback with phase
-        progress_callback = None
-        if self.progress_callback and progress_phase:
-
-            def progress_callback(current: int, total: int) -> None:
-                # Handle both sync and async callbacks
-                assert self.progress_callback is not None  # Guaranteed by outer condition
-                callback_result = self.progress_callback(progress_phase, current, total)
-                if callback_result is not None:
-                    # Track the task to ensure it's awaited before the pipeline finishes
-                    async def _wrap():
-                        await callback_result
-
-                    task = asyncio.create_task(_wrap())
-                    self._pending_progress_tasks.append(task)
-
-        from src.core.config import config
-
-        return ClauseaCrawler(
-            max_depth=max_depth or self.max_depth,
-            max_pages=max_pages or self.max_pages,
-            max_concurrent=self.concurrent_limit,
-            delay_between_requests=self.delay_between_requests,
-            delay_jitter=config.crawler.rate_limit_jitter,
-            timeout=self.timeout,
-            allowed_domains=self._allowed_domains_for_product(product),
-            respect_robots_txt=self.respect_robots_txt,
-            user_agent=self.user_agent,
-            follow_external_links=False,
-            min_legal_score=min_legal_score if min_legal_score is not None else 0.0,
-            strategy=strategy or self.crawler_strategy,
-            log_file_path=log_file_path,
-            use_browser=self.use_browser,
-            browser_concurrency=self.browser_concurrency,
-            proxy=self.proxy,
-            allowed_paths=product.crawl_allowed_paths,
-            denied_paths=product.crawl_denied_paths,
-            progress_callback=progress_callback,
-            result_callback=result_callback,
-            stop_callback=stop_callback,
-            recently_stored_urls=recently_stored_urls,
-        )
-
-    async def _store_documents(self, documents: list[Document]) -> int:
-        """
-        Store documents with intelligent deduplication and update logic.
-
-        Args:
-            documents: List of documents to store
-
-        Returns:
-            Number of documents actually stored (new + updated)
-        """
-        stored_count = 0
-        linked_count = 0
-        updated_count = 0
-        duplicate_count = 0
-        error_count = 0
-
-        # Process canonical (locale-neutral, shorter) URLs first so they win as the
-        # representative when collapsing same-content locale variants.
-        documents = sorted(documents, key=lambda d: _canonical_rank(d.url))
-        # (product_id, content fingerprint) -> URL kept for that content. Lets us collapse
-        # locale variants that serve identical policy text (e.g. /privacy and /es/privacy)
-        # while keeping genuinely different regional policies (different text → different fp).
-        seen_fingerprints: dict[tuple[str | None, str], str] = {}
-
-        async with db_session() as db:
-            document_service = create_document_service()
-
-            for document in documents:
-                try:
-                    source_product_id = document.product_id
-                    # Canonical URL lookup: the same policy URL can be reused by many products.
-                    existing_doc = await document_service.get_document_by_url(db, document.url)
-                    if existing_doc:
-                        linked_this_run = False
-                        if not existing_doc.is_linked_to_product(source_product_id):
-                            linked_this_run = await document_service.link_document_to_product(
-                                db, existing_doc.id, source_product_id
-                            )
-                            if linked_this_run:
-                                logger_storage.info(
-                                    "linked existing canonical document %s to product %s",
-                                    document.url,
-                                    source_product_id,
-                                )
-
-                        # Calculate content + metadata hashes for comparison
-                        # We include metadata that the LLM might have updated (title, doc_type, locale, regions, effective_date)
-                        # We use separators to avoid potential string collisions
-                        metadata_str = f"|{document.title}|{document.doc_type}|{document.locale}|{','.join(document.regions)}|{document.effective_date}|"
-                        current_hash = hashlib.sha256(
-                            (document.text + metadata_str).encode()
-                        ).hexdigest()
-
-                        existing_metadata_str = f"|{existing_doc.title}|{existing_doc.doc_type}|{existing_doc.locale}|{','.join(existing_doc.regions)}|{existing_doc.effective_date}|"
-                        existing_hash = hashlib.sha256(
-                            (existing_doc.text + existing_metadata_str).encode()
-                        ).hexdigest()
-
-                        if current_hash != existing_hash:
-                            # Never overwrite a document that has content with an empty-content
-                            # version. This can happen when a JS-rendered page is fetched by
-                            # the static crawler on a subsequent run: the shell HTML passes the
-                            # raw-text sufficiency check but yields no extracted body, so the
-                            # pipeline creates a Document with text="" that would silently erase
-                            # previously stored policy text.
-                            if not document.text.strip() and existing_doc.text.strip():
-                                logger_storage.warning(
-                                    f"refusing to overwrite non-empty document with empty content: {document.url}"
-                                )
-                                self.stats.duplicates_skipped += 1
-                                duplicate_count += 1
-                                continue
-
-                            # Keep canonical owner stable while preserving all linked products.
-                            document.product_id = existing_doc.product_id
-                            document.product_ids = [
-                                *existing_doc.product_ids,
-                                *(
-                                    [source_product_id]
-                                    if source_product_id not in existing_doc.product_ids
-                                    else []
-                                ),
-                            ]
-                            changed = _diff_fields(existing_doc, document)
-                            await DocumentVersionRepository().archive(
-                                db, existing_doc, job_id=self._job_id, changed_fields=changed
-                            )
-
-                            # Update existing document with new content/metadata
-                            logger_storage.info(
-                                f"updating existing document with changes: {document.url}"
-                            )
-                            document.id = existing_doc.id  # Preserve original ID
-                            document.content_hash = _content_fingerprint(document.text)
-                            await document_service.update_document(db, document)
-                            stored_count += 1
-                            updated_count += 1
-                        else:
-                            if linked_this_run:
-                                # Reused canonical doc for a new product: count this as a
-                                # successful stored document for this product run so synthesis
-                                # can proceed even when no new canonical rows are inserted.
-                                stored_count += 1
-                                linked_count += 1
-                            else:
-                                logger_storage.debug(
-                                    f"skipping unchanged document (duplicate): {document.url}"
-                                )
-                                self.stats.duplicates_skipped += 1
-                                duplicate_count += 1
-                        if document.text and document.text.strip():
-                            seen_fingerprints[
-                                (source_product_id, _content_fingerprint(document.text))
-                            ] = document.url
-                    else:
-                        # Collapse same-content locale variants: if another doc for this
-                        # product already holds identical policy text at a different URL,
-                        # skip this one. Different regional policies have different text and
-                        # therefore a different fingerprint, so they are never merged.
-                        if document.text and document.text.strip():
-                            fp_key = (
-                                source_product_id,
-                                _content_fingerprint(document.text),
-                            )
-                            kept_url = seen_fingerprints.get(fp_key)
-                            if kept_url and kept_url != document.url:
-                                logger_storage.info(
-                                    f"collapsing same-content variant {document.url} "
-                                    f"(identical to {kept_url})"
-                                )
-                                self.stats.duplicates_skipped += 1
-                                duplicate_count += 1
-                                continue
-                            seen_fingerprints[fp_key] = document.url
-
-                        # Create new document
-                        logger_storage.info(f"storing new document: {document.url}")
-                        document.content_hash = _content_fingerprint(document.text)
-                        await document_service.store_document(db, document)
-                        stored_count += 1
-
-                except Exception as e:
-                    logger_storage.error(
-                        f"failed to store document {document.url}: {e}", exc_info=True
-                    )
-                    error_count += 1
-
-        # Log summary of storage operation
-        if len(documents) > 0:
-            new_count = stored_count - updated_count - linked_count
-            logger_storage.info(
-                f"storage complete: {stored_count} stored "
-                f"({new_count} new, {updated_count} updated, {linked_count} linked), "
-                f"{duplicate_count} duplicates skipped, {error_count} errors"
-            )
-
-        return stored_count
-
-    async def _process_crawl_result(
-        self, result: CrawlResult, product: Product, trusted: bool = False
-    ) -> Document | None:
-        """
-        Process a single crawl result through the analysis pipeline.
-
-        Args:
-            result: Crawl result from ClauseaCrawler
-            product: Product being processed
-
-        Returns:
-            Document if legal and English, None otherwise
-        """
-        self.analyzer.reset_usage_stats()
-        usage_reason = "completed"
-        document: Document | None = None
-
-        try:
-            text_content = markdown_to_text(result.markdown)
-
-            # Skip empty or very short content without wasting LLM calls
-            if not text_content or len(text_content.strip()) < 300:
-                text_len = len(text_content.strip()) if text_content else 0
-                self.stats.crawl_skip_reasons.append(
-                    {
-                        "url": result.url,
-                        "reason": "insufficient_content",
-                        "detail": f"text={text_len} chars",
-                    }
-                )
-                logger_analysis.info(
-                    f"[skip:insufficient_content] {result.url} (text={text_len} chars)"
-                )
-                return None
-
-            # Skip garbled/binary content that slipped through
-            if ClauseaCrawler._is_garbled_content(text_content):
-                self.stats.crawl_skip_reasons.append(
-                    {"url": result.url, "reason": "garbled_content", "detail": None}
-                )
-                logger_analysis.info(f"[skip:garbled_content] {result.url}")
-                return None
-
-            if (
-                not trusted
-                and result.legal_score is not None
-                and result.legal_score < MIN_LEGAL_SCORE_THRESHOLD
-            ):
-                self.stats.crawl_skip_reasons.append(
-                    {
-                        "url": result.url,
-                        "reason": "low_legal_score",
-                        "detail": f"legal_score={result.legal_score:.2f}",
-                    }
-                )
-                logger_analysis.info(
-                    f"[skip:low_legal_score] {result.url} (legal_score={result.legal_score:.2f})"
-                )
-                return None
-
-            # Always classify before storing; skip non-policy/legal pages and get proper title
-            classification = await self.analyzer.classify_document(
-                result.url, text_content, result.metadata
-            )
-
-            logger_analysis.debug(
-                f"classified as '{classification.get('classification')}' (is_policy: {classification.get('is_policy_document')})"
-            )
-
-            # Skip pages that are not policy documents
-            if not classification.get("is_policy_document", False):
-                classification_label = str(classification.get("classification", "unknown"))
-                self.stats.crawl_skip_reasons.append(
-                    {
-                        "url": result.url,
-                        "reason": "non_policy_classification",
-                        "detail": f"classifier={classification_label}",
-                    }
-                )
-                logger_analysis.info(
-                    f"[skip:non_policy_classification] {result.url} (classifier={classification_label})"
-                )
-                usage_reason = f"non-policy classification: {classification_label}"
-                return None
-
-            # Detect locale only for policy documents
-            locale_result = await self.analyzer.detect_locale(
-                text_content, result.metadata, result.url
-            )
-            detected_locale = locale_result.get("locale", "en-US")
-            language_name = locale_result.get("language_name", "English")
-
-            logger_analysis.debug(
-                f"detected locale: {detected_locale} ({language_name}, confidence: {locale_result.get('confidence', 0):.2f})"
-            )
-
-            # Skip non-English documents*
-            # TODO: Support other languages
-            if "en" not in detected_locale.lower():
-                self.stats.crawl_skip_reasons.append(
-                    {
-                        "url": result.url,
-                        "reason": "non_english",
-                        "detail": f"locale={detected_locale}",
-                    }
-                )
-                logger_analysis.info(f"[skip:non_english] {result.url} (locale={detected_locale})")
-                self.stats.non_english_skipped += 1
-                usage_reason = f"non-English locale: {detected_locale}"
-                return None
-
-            self.stats.english_documents += 1
-            self.stats.policy_documents_processed += 1
-
-            # Detect regions
-            region_detection = await self.analyzer.detect_regions(
-                text_content, result.metadata, result.url
-            )
-
-            # Extract effective date
-            effective_date_str = await self.analyzer.extract_effective_date(
-                text_content, result.metadata
-            )
-            effective_date = None
-            if effective_date_str:
-                try:
-                    effective_date = datetime.strptime(effective_date_str, "%Y-%m-%d")
-                    logger.debug(f"Parsed effective date: {effective_date}")
-                except ValueError as e:
-                    logger.warning(f"Failed to parse effective date '{effective_date_str}': {e}")
-
-            # Extract title
-            doc_type = coerce_doc_type_from_classifier(classification.get("classification"))
-            title_result = await self.analyzer.extract_title(
-                result.markdown,
-                result.metadata,
-                result.url,
-                doc_type,
-            )
-
-            # Create document
-            document = Document(
-                title=title_result.get("title", "Untitled Policy Document"),
-                url=result.url,
-                product_id=product.id,
-                markdown=result.markdown,
-                text=text_content,
-                metadata=result.metadata,
-                doc_type=doc_type,
-                locale=detected_locale,
-                regions=region_detection.get("regions", ["global"]),
-                effective_date=effective_date,
-            )
-
-            effective_date_info = (
-                f", effective: {document.effective_date.strftime('%Y-%m-%d')}"
-                if document.effective_date
-                else ""
-            )
-            logger_analysis.info(
-                f"analyzed document: '{document.title}' ({document.doc_type}, {document.locale}, {document.regions}{effective_date_info})"
-            )
-
-            usage_reason = "success"
-            return document
-
-        except Exception as e:
-            usage_reason = f"error: {e.__class__.__name__}"
-            logger.error(f"Failed to process crawl result {result.url}: {e}")
-            return None
-        finally:
-            # Extract document info if available (document may not exist if processing failed)
-            document_id = document.id if document else None
-            document_title = document.title if document else None
-
-            # Get usage summary before it's consumed by log_llm_usage
-            usage_summary = self.analyzer.get_usage_summary()
-
-            # Aggregate usage stats for pipeline totals
-            for model_stats in usage_summary.values():
-                self.stats.total_prompt_tokens += model_stats.get("prompt_tokens", 0)
-                self.stats.total_completion_tokens += model_stats.get("completion_tokens", 0)
-                self.stats.total_tokens += model_stats.get("total_tokens", 0)
-                cost = model_stats.get("cost")
-                if cost is not None and cost > 0:
-                    self.stats.total_cost += cost
-
-            # Now log the usage (this will consume the summary)
-            self.analyzer.log_llm_usage(
-                context=result.url,
-                reason=usage_reason,
-                operation_type="crawl",
-                product_slug=product.slug,
-                product_id=product.id,
-                document_url=result.url,
-                document_title=document_title,
-                document_id=document_id,
-            )
-
-    def _normalize_url(self, url_or_domain: str) -> str:
-        """
-        Normalize a URL or domain to ensure it has a protocol.
-
-        Args:
-            url_or_domain: URL or domain string
-
-        Returns:
-            Normalized URL with https:// protocol
-        """
-        url_or_domain = url_or_domain.strip()
-        if not url_or_domain:
-            return url_or_domain
-
-        # If already has a protocol, use it as-is
-        if url_or_domain.startswith(("http://", "https://")):
-            return url_or_domain
-
-        # Prepend https://
-        return f"https://{url_or_domain}"
-
-    def _seed_dedupe_key(self, normalized_url: str) -> tuple[str, str, str, str]:
-        """Stable dedupe key for crawl seeds.
-
-        Treat root variants as equivalent (``https://x.com`` == ``https://x.com/``)
-        while preserving deeper paths and query-bearing override seeds.
-        """
-        parsed = urlparse(normalized_url)
-        path = parsed.path
-        if path in ("", "/"):
-            path = ""
-        return (
-            parsed.scheme.lower(),
-            parsed.netloc.lower(),
-            path,
-            parsed.query,
-        )
-
-    def _get_crawl_urls(self, product: Product) -> list[str]:
-        """
-        Get crawl URLs for a product with domain roots as primary seeds.
-
-        Args:
-            product: Product to get URLs for
-
-        Returns:
-            List of URLs to crawl (all normalized with https:// protocol)
-        """
-        # Domain roots are the backbone (always crawl homepage + sitemap discovery).
-        candidates: list[str] = []
-        for domain in product.domains:
-            normalized = self._normalize_url(domain)
-            if normalized:
-                candidates.append(normalized)
-
-        # crawl_base_urls are optional overrides for odd sites, not the primary source.
-        for seed in product.crawl_base_urls or []:
-            normalized = self._normalize_url(seed)
-            if normalized:
-                candidates.append(normalized)
-
-        deduped: list[str] = []
-        seen: set[tuple[str, str, str, str]] = set()
-        for candidate in candidates:
-            key = self._seed_dedupe_key(candidate)
-            if key in seen:
-                continue
-            seen.add(key)
-            deduped.append(candidate)
-        return deduped
-
-    async def _start_crawl_session(self, product: Product, crawl_urls: list[str]) -> CrawlSession:
-        session = CrawlSession(
-            product_id=product.id,
-            product_slug=product.slug,
-            seed_urls=crawl_urls,
-            status="running",
-            settings={
-                "max_depth": self.max_depth,
-                "max_pages": self.max_pages,
-                "discovery_max_depth": self.discovery_max_depth,
-                "discovery_max_pages": self.discovery_max_pages,
-                "discovery_strategy": self.discovery_strategy,
-                "discovery_min_legal_score": self.discovery_min_legal_score,
-                "fallback_strategy": self.fallback_strategy,
-                "default_strategy": self.crawler_strategy,
-            },
-            stats={"mode": "hybrid"},
-        )
-        async with db_session() as db:
-            repo = CrawlRepository()
-            await repo.create_session(db, session)
-        return session
-
-    async def _finish_crawl_session(
-        self, session: CrawlSession, stats: dict[str, int], error: str | None = None
-    ) -> None:
-        session.stats.update(stats)
-        session.status = "failed" if error else "completed"
-        session.error = error
-        session.completed_at = datetime.now()
-        async with db_session() as db:
-            repo = CrawlRepository()
-            await repo.update_session(db, session)
-
-    def _should_fallback_crawl(self, documents: list[Document]) -> bool:
-        if len(documents) < self.min_docs_before_fallback:
-            return True
-        doc_types = {doc.doc_type for doc in documents}
-        return not all(req in doc_types for req in self.required_doc_types)
-
-    async def _classify_results(
-        self,
-        results: list[CrawlResult],
-        product: Product,
-        processed_urls: set[str],
-        seen_fingerprints: set[str],
-        trusted_urls: frozenset[str] | None = None,
-    ) -> list[Document]:
-        """Classify a batch of crawl results and return policy documents.
-
-        Mutates *processed_urls* to track deduplication across calls.
-        Mutates *seen_fingerprints* for near-duplicate content skipping across passes.
-        Failed crawl results are collected in ``self.stats.crawl_errors``.
-        trusted_urls: URLs that bypass the legal_score gate (seed URLs whose path
-        carries no policy keywords despite containing policy content).
-        """
-        documents: list[Document] = []
-        for result in results:
-            if result.url in processed_urls:
-                continue
-            processed_urls.add(result.url)
-            if result.success:
-                fp = _content_fingerprint(result.content or "")
-                if fp in seen_fingerprints:
-                    logger.debug(f"Skipping near-duplicate: {result.url}")
-                    continue
-                seen_fingerprints.add(fp)
-                is_trusted = trusted_urls is not None and result.url in trusted_urls
-                document = await self._process_crawl_result(result, product, trusted=is_trusted)
-                if document:
-                    documents.append(document)
-            else:
-                logger.warning(f"Failed to crawl {result.url}: {result.error_message}")
-                self.stats.crawl_errors.append(
-                    {
-                        "url": result.url,
-                        "status_code": result.status_code,
-                        "error_message": result.error_message,
-                        "error_type": classify_crawl_error(
-                            result.error_message, result.status_code
-                        ),
-                    }
-                )
-        return documents
-
-    async def _crawl_base_urls(
-        self, crawler: ClauseaCrawler, crawl_urls: list[str]
-    ) -> list[CrawlResult]:
-        results: list[CrawlResult] = []
-        logger.info(f"Starting crawl across {len(crawl_urls)} seed URL(s): {', '.join(crawl_urls)}")
-        results = await crawler.crawl_multiple(crawl_urls)
-        logger.info(
-            f"Finished crawl across {len(crawl_urls)} seed URL(s): "
-            f"{len(results)} result(s) returned"
-        )
-        return results
-
-    async def _process_product(self, product: Product) -> list[Document]:
-        """
-        Process a single product through the complete pipeline.
-
-        Args:
-            product: Product to process
-
-        Returns:
-            List of processed and stored documents
-        """
-        product_start_time = time.time()
-        log_memory_usage(f"Starting {product.name}")
-
-        # Get crawl URLs (domain roots primary, crawl_base_urls as optional overrides).
-        crawl_urls = self._get_crawl_urls(product)
-
-        if not crawl_urls:
-            logger.warning(
-                f"No crawl base URLs or domains for {product.name}. "
-                f"Cannot crawl without starting URLs."
-            )
-            self.stats.products_failed += 1
-            self.stats.failed_product_slugs.append(product.slug)
-            return []
-
-        has_domains = bool(product.domains)
-        has_overrides = bool(product.crawl_base_urls)
-        if has_domains and has_overrides:
-            source = "domains+overrides"
-        elif has_domains:
-            source = "domains"
-        else:
-            source = "crawl_base_urls_override"
-        crawl_session: CrawlSession | None = None
-        try:
-            logger.info(
-                f"🕷️ Crawling {product.name} ({len(product.domains)} domains) "
-                f"from {len(crawl_urls)} base URLs (using {source})"
-            )
-
-            # Create crawl session for auditability
-            crawl_session = await self._start_crawl_session(product, crawl_urls)
-
-            processed_urls: set[str] = set()
-            seen_fingerprints: set[str] = set()
-            total_results = 0
-            processed_documents: list[Document] = []
-
-            # crawl_base_urls are explicit overrides provided by the user or extension.
-            # Unlike domain roots (derived from product.domains), these are known policy
-            # pages and bypass the URL-score gate in _process_crawl_result — their path
-            # may carry no policy keywords (e.g. amazon's /gp/help/customer/display.html)
-            # even though the content is clearly a policy document. Domain roots still go
-            # through normal scoring. The content-based LLM classifier runs for all.
-            trusted_urls: frozenset[str] = frozenset(
-                self._normalize_url(u)
-                for u in (product.crawl_base_urls or [])
-                if self._normalize_url(u)
-            )
-
-            # Stream each crawled result through classification + storage as it is
-            # produced, so a crawl killed at the time ceiling or by a stall keeps the
-            # work done so far instead of losing everything in an end-of-crawl batch.
-            # The shared *processed_urls* / *seen_fingerprints* sets dedupe across the
-            # per-result calls exactly as the old batch call did. _store_documents is
-            # idempotent (dedupes by URL/fingerprint), so re-arrivals are safe.
-            async def result_callback(result: CrawlResult) -> None:
-                docs = await self._classify_results(
-                    [result],
-                    product,
-                    processed_urls,
-                    seen_fingerprints,
-                    trusted_urls=trusted_urls,
-                )
-                if docs:
-                    stored = await self._store_documents(docs)
-                    self.stats.policy_documents_stored += stored
-                    processed_documents.extend(docs)
-
-            # Docs stored within the resume freshness window. A retried crawl skips
-            # re-fetching these so it resumes instead of re-rendering hours of work;
-            # docs older than the window are still re-fetched so scheduled monitoring
-            # re-crawls days later keep detecting policy changes. One query per crawl.
-            recently_stored_urls = await self._get_recently_stored_urls(product)
-            if recently_stored_urls:
-                logger.info(
-                    f"♻️ Resume: skipping re-fetch of {len(recently_stored_urls)} "
-                    f"recently stored docs for {product.name} "
-                    f"(within {RESUME_FRESH_HOURS}h window)"
-                )
-
-            # ----------------------------------------------------------
-            # Stage 1: Crawl (streams classify + store per page)
-            # ----------------------------------------------------------
-
-            # Discovery pass (precision-first)
-            logger_discovery.info(
-                f"starting discovery pass for '{product.name}' (max_depth={self.discovery_max_depth}, max_pages={self.discovery_max_pages})"
-            )
-            discovery_crawler = self._create_crawler_for_product(
-                product,
-                max_depth=self.discovery_max_depth,
-                max_pages=self.discovery_max_pages,
-                min_legal_score=self.discovery_min_legal_score,
-                strategy=self.discovery_strategy,
-                progress_phase="discovery",
-                result_callback=result_callback,
-                recently_stored_urls=recently_stored_urls,
-            )
-            discovery_results = await self._crawl_base_urls(discovery_crawler, crawl_urls)
-            logger_discovery.info(
-                f"discovery pass complete for '{product.name}': found {len(discovery_results)} pages"
-            )
-            total_results += len(discovery_results)
-
-            # Fallback deep crawl (recall-first) if coverage is low. The decision runs
-            # on the documents already streamed+stored during the discovery pass.
-            if self._should_fallback_crawl(processed_documents):
-                logger_discovery.info(
-                    f"discovery coverage insufficient for '{product.name}'; starting fallback deep crawl (max_depth={self.max_depth}, max_pages={self.max_pages})"
-                )
-                fallback_crawler = self._create_crawler_for_product(
-                    product,
-                    max_depth=self.max_depth,
-                    max_pages=self.max_pages,
-                    min_legal_score=self.fallback_min_legal_score,
-                    strategy=self.fallback_strategy,
-                    progress_phase="fallback",
-                    result_callback=result_callback,
-                    recently_stored_urls=recently_stored_urls,
-                )
-                fallback_results = await self._crawl_base_urls(fallback_crawler, crawl_urls)
-                logger_discovery.info(
-                    f"fallback pass complete for '{product.name}': found {len(fallback_results)} pages"
-                )
-                total_results += len(fallback_results)
-
-            self.stats.total_urls_crawled += total_results
-            self.stats.total_documents_found += total_results
-
-            logger.info(f"📄 Found {total_results} pages for {product.name}")
-
-            if processed_documents:
-                logger.info(
-                    f"💾 Stored {len(processed_documents)} policy documents for {product.name}"
-                )
-            else:
-                logger.info(f"No policy documents found for {product.name}")
-
-            await self._finish_crawl_session(
-                crawl_session,
-                stats={
-                    "pages_crawled": total_results,
-                    "policy_documents": len(processed_documents),
-                },
-            )
-
-            self.stats.products_processed += 1
-
-            product_duration = time.time() - product_start_time
-            log_memory_usage(f"Completed {product.name}")
-            logger.info(
-                f"✅ Completed {product.name} in {product_duration:.2f}s "
-                f"({len(processed_documents)} policy docs this run)"
-            )
-
-            return processed_documents
-
-        except Exception as e:
-            logger.error(f"Failed to process product {product.name}: {e}")
-            if crawl_session:
-                await self._finish_crawl_session(
-                    crawl_session, stats={"pages_crawled": 0, "policy_documents": 0}, error=str(e)
-                )
-            self.stats.products_failed += 1
-            self.stats.failed_product_slugs.append(product.slug)
-            return []
-
-    async def run(self, products: list[Product] | None = None) -> ProcessingStats:
-        """
-        Execute the complete policy document crawling pipeline.
-
-        Returns:
-            ProcessingStats with comprehensive pipeline metrics
-        """
-        # Start comprehensive monitoring
-        tracemalloc.start()
-        pipeline_start_time = time.time()
-
-        logger.info("🚀 Starting Policy Document Crawling Pipeline")
-        log_memory_usage("Pipeline start")
-
-        # Start background memory monitoring
-        memory_task = asyncio.create_task(memory_monitor_task(60))
-
-        try:
-            # Get all products
-            if products is None:
-                async with db_session() as db:
-                    product_service = create_product_service()
-                    products = await product_service.get_all_products(db)
-            logger.info(f"📊 Processing {len(products)} products")
-
-            # Use a semaphore to limit parallel products
-            semaphore = asyncio.Semaphore(self.max_parallel_products)
-
-            async def _process_product_with_semaphore(idx: int, product: Product) -> None:
-                async with semaphore:
-                    logger.info(f"🏢 [{idx}/{len(products)}] Starting product: {product.name}")
-                    await self._process_product(product)
-                    logger.info(f"✅ [{idx}/{len(products)}] Finished product: {product.name}")
-
-            # Create tasks for all products
-            tasks = [
-                _process_product_with_semaphore(i, product) for i, product in enumerate(products, 1)
-            ]
-
-            # Execute tasks in parallel with limited concurrency
-            await asyncio.gather(*tasks)
-
-            # Ensure all progress callbacks are finished before returning
-            if self._pending_progress_tasks:
-                await asyncio.gather(*self._pending_progress_tasks, return_exceptions=True)
-                self._pending_progress_tasks.clear()
-
-            # Calculate final statistics
-            self.stats.processing_time_seconds = time.time() - pipeline_start_time
-
-            # Log comprehensive results
-            logger.info("🎉 Pipeline completed successfully!")
-            logger.info(f"📊 Products processed: {self.stats.products_processed}")
-            logger.info(f"❌ Products failed: {self.stats.products_failed}")
-            if self.stats.failed_product_slugs:
-                logger.info(
-                    f"❌ Failed product slugs: {', '.join(self.stats.failed_product_slugs)}"
-                )
-            logger.info(f"🌐 Total URLs crawled: {self.stats.total_urls_crawled}")
-            logger.info(f"📄 Total documents found: {self.stats.total_documents_found}")
-            logger.info(f"📋 Policy documents processed: {self.stats.policy_documents_processed}")
-            logger.info(f"💾 Policy documents stored: {self.stats.policy_documents_stored}")
-            logger.info(f"🗣️ English documents: {self.stats.english_documents}")
-            logger.info(f"🌍 Non-English skipped: {self.stats.non_english_skipped}")
-            logger.info(f"🔄 Duplicates skipped: {self.stats.duplicates_skipped}")
-            logger.info(f"✅ Success rate: {self.stats.success_rate:.1f}%")
-            logger.info(f"🎯 Legal detection rate: {self.stats.legal_detection_rate:.1f}%")
-            # Format time as minutes or hours + minutes
-            total_seconds = self.stats.processing_time_seconds
-            if total_seconds >= 3600:
-                hours = int(total_seconds // 3600)
-                minutes = int((total_seconds % 3600) // 60)
-                time_str = f"{hours}h {minutes}m"
-            else:
-                minutes = int(total_seconds // 60)
-                seconds = int(total_seconds % 60)
-                time_str = f"{minutes}m {seconds}s" if minutes > 0 else f"{seconds}s"
-            logger.info(f"⏱️ Total time: {time_str}")
-
-            # Log LLM token usage and cost
-            if self.stats.total_tokens > 0:
-                cost_str = f" (${self.stats.total_cost:.6f})" if self.stats.total_cost > 0 else ""
-                logger.info(
-                    f"🔢 LLM tokens: input={self.stats.total_prompt_tokens:,} "
-                    f"output={self.stats.total_completion_tokens:,} "
-                    f"total={self.stats.total_tokens:,}{cost_str}"
-                )
-
-            return self.stats
-
-        finally:
-            # Cleanup and final monitoring
-            memory_task.cancel()
-            log_memory_usage("Pipeline end")
-
-            current, peak = tracemalloc.get_traced_memory()
-            logger.info(
-                f"🧠 Memory usage: Current={current / 1024 / 1024:.1f}MB, "
-                f"Peak={peak / 1024 / 1024:.1f}MB"
-            )
-            tracemalloc.stop()
-
-
-async def main() -> None:
-    """Main entry point for the crawling pipeline."""
-    try:
-        pipeline = PolicyDocumentPipeline()
-
-        stats = await pipeline.run()
-
-        # Exit with appropriate code
-        if stats.products_failed > 0:
-            failed_slugs = (
-                ", ".join(stats.failed_product_slugs) if stats.failed_product_slugs else "unknown"
-            )
-            logger.warning(
-                f"Pipeline completed with {stats.products_failed} failures: {failed_slugs}"
-            )
-        else:
-            logger.info("Pipeline completed successfully")
-        exit(0)
-
-    except KeyboardInterrupt:
-        logger.warning("Pipeline interrupted by user")
-        exit(130)
-    except Exception as e:
-        logger.error(f"Pipeline failed with error: {e}")
-        exit(1)
-
-
-if __name__ == "__main__":
-    asyncio.run(main())
